+++ src/Shared/Modules/Aero/ClientEvent.lua
@@ -1,0 +1,10 @@
+local function ClientEvent(wrappedEvent)
+    return {
+        _isServer = false,
+        _isEvent = true,
+        _isAllClients = false,
+        _event = wrappedEvent,
+    }
+end
+
+return ClientEvent