+++ src/Shared/Modules/Aero/Controller.lua
@@ -2,18 +2,6 @@ local Controller = {}
 
 -- Members
 Controller.__index = {
-    ConnectEvent = function(self, name, listener)
-    end,
-    RegisterEvent = function(self, name)
-    end,
-    FireEvent = function(self, name)
-    end,
-    FireClientEvent = function(self, name, player, ...)
-        local args = { ... }
-    end,
-    FireAllClientsEvent = function(self, name, ...)
-        local args = { ... }
-    end,
 };
 
 -- Statics