+++ src/Shared/Modules/Aero/ServerAsync.lua
@@ -1,0 +1,10 @@
+local function ServerAsync(wrappedFunction)
+    return {
+        _isServer = true,
+        _isAsync = true,
+        _isVoid = false,
+        _func = wrappedFunction,
+    }
+end
+
+return ServerAsync