+++ src/Shared/Modules/Aero/ServerAsyncVoid.lua
@@ -1,0 +1,10 @@
+local function ServerAsyncVoid(wrappedFunction)
+    return {
+        _isServer = true,
+        _isAsync = true,
+        _isVoid = true,
+        _func = wrappedFunction,
+    }
+end
+
+return ServerAsyncVoid